move setup_coverage to testing
where it belongs
