+++ buildout.cfg
@@ -29,7 +29,7 @@ eggs =
   zc.recipe.egg
   coverage
 initialization =
-  from zc.buildout import setup_coverage
+  from zc.buildout.testing import setup_coverage
   setup_coverage()
 
 # Tests that can be run wo a network
